requirements.txt
streamlit
pillow
pytesseract
selenium
pytz
